chore: rename <LoadingList/> to <LoadingSearchResults/>
